testdata/requirement.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
bitarray==2.9.2
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
ckzg==1.0.2
cytoolz==0.12.3
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
hexbytes==0.3.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
jsonschema==4.17.3
lru-dict==1.2.0
multidict==6.0.5
numpy==1.21.6
packaging==23.1
pandas==1.3.5
parsimonious==0.9.0
pkgutil_resolve_name==1.3.10
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2024.2
pyunormalize==16.0.0
regex==2024.4.16
requests==2.31.0
rlp==3.0.0
rsa==4.9
sh==1.14.3
six==1.16.0
toolz==0.12.1
tqdm==4.66.5
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
web3==6.20.3
websockets==11.0.3
yarl==1.9.4
zipp==3.15.0
